requirements.txt
flask
python-docx
werkzeug
PyPDF2